짝수와 홀수 풀이
